LayoutTests/fast/visual-viewport/zoomed-scroll-into-view-fixed-expected.txt

Tests revealing elements inside position:fixed after zooming.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Reveal "left-target"
PASS document.scrollingElement.scrollTop is 838
PASS document.scrollingElement.scrollLeft is 40

Reveal "bottom-target"
PASS document.scrollingElement.scrollTop is 1048
PASS document.scrollingElement.scrollLeft is 40

Reveal "right-target"
PASS document.scrollingElement.scrollTop is 1086
PASS document.scrollingElement.scrollLeft is 333

Reveal "top-target"
PASS document.scrollingElement.scrollTop is 834
PASS document.scrollingElement.scrollLeft is 230
PASS successfullyParsed is true

TEST COMPLETE
